pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-trading-bot"
version = "0.1.0"
description = "ML Trading Bot"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "ta>=0.10.2",
    "joblib>=1.3.0",
    "pyyaml>=6.0",
    "pytest>=7.0.0",
    "jupyter>=1.0.0",
    "inquirer>=3.2.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "lightstreamer-client-lib>=1.1.4",
    "alpaca-py>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "jupyter>=1.0.0",
]

[project.scripts]
trading-bot = "src.menu:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v"



